Revert debug hooks code move
