+++ worker/uniter/runner/debug/client_test.go
@@ -9,7 +9,7 @@ import (
 
 	gc "gopkg.in/check.v1"
 
-	"github.com/juju/juju/worker/common/runner/debug"
+	"github.com/juju/juju/worker/uniter/runner/debug"
 )
 
 type DebugHooksClientSuite struct{}
